src/app/api/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
strawberry-graphql==0.216.1
requests==2.31.0
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.0
